Move tests for time class into separate file
