+++ test/tests/Suite.ts
@@ -116,6 +116,7 @@ import "./calendar/CalendarModelTest.js"
 import "./calendar/CalendarEventUpdateCoordinatorTest.js"
 import "./calendar/CalendarUtilsTest.js"
 import "./calendar/CalendarViewModelTest.js"
+import "./calendar/date/TimeTest.js"
 import "./calendar/EventDragHandlerTest.js"
 import "./calendar/eventeditor/CalendarEventAlarmModelTest.js"
 import "./calendar/eventeditor/CalendarEventModelTest.js"
