+++ test/tests/calendar/CalendarUtilsTest.ts
@@ -48,7 +48,6 @@ import { createTestEntity } from "../TestUtils.js"
 import { getCalendarMonth, getEventType } from "../../../src/applications/calendar-app/calendar/gui/CalendarGuiUtils.js"
 import { EventType } from "../../../src/applications/calendar-app/calendar/gui/eventeditor-model/CalendarEventModel.js"
 import { CalendarInfo } from "../../../src/applications/calendar-app/calendar/model/CalendarModel.js"
-import { Time } from "../../../src/applications/common/calendar/date/Time.js"
 import type { UserController } from "../../../src/applications/common/api/main/UserController.js"
 import { EventWrapper } from "../../../src/applications/calendar-app/calendar/view/CalendarViewModel"
 import {
@@ -428,344 +427,6 @@ o.spec("CalendarUtilsTest", function () {
 					"27,28,29,30,31,1,2\n" +
 					"3,4,5,6,7,8,9",
 			)
-		})
-	})
-	o.spec("Time Class", function () {
-		o.spec("parseTimeTo", function () {
-			function parseTimeString(timeString: string): { hours: number; minutes: number } {
-				return neverNull(Time.parseFromString(timeString)?.toObject() ?? null)
-			}
-
-			o("parses full 24H time", function () {
-				o(parseTimeString("12:45")).deepEquals({
-					hours: 12,
-					minutes: 45,
-				})
-				o(parseTimeString("1245")).deepEquals({
-					hours: 12,
-					minutes: 45,
-				})
-				o(parseTimeString("2359")).deepEquals({
-					hours: 23,
-					minutes: 59,
-				})
-				o(parseTimeString("0000")).deepEquals({
-					hours: 0,
-					minutes: 0,
-				})
-				o(parseTimeString("0623")).deepEquals({
-					hours: 6,
-					minutes: 23,
-				})
-				o(parseTimeString("08:09")).deepEquals({
-					hours: 8,
-					minutes: 9,
-				})
-			})
-			o("parses partial 24H time", function () {
-				o(parseTimeString("12")).deepEquals({
-					hours: 12,
-					minutes: 0,
-				})
-				o(parseTimeString("1:2")).deepEquals({
-					hours: 1,
-					minutes: 2,
-				})
-				o(parseTimeString("102")).deepEquals({
-					hours: 1,
-					minutes: 2,
-				})
-				o(parseTimeString("17")).deepEquals({
-					hours: 17,
-					minutes: 0,
-				})
-				o(parseTimeString("6")).deepEquals({
-					hours: 6,
-					minutes: 0,
-				})
-				o(parseTimeString("955")).deepEquals({
-					hours: 9,
-					minutes: 55,
-				})
-				o(parseTimeString("12:3")).deepEquals({
-					hours: 12,
-					minutes: 3,
-				})
-				o(parseTimeString("809")).deepEquals({
-					hours: 8,
-					minutes: 9,
-				})
-			})
-			o("not parses incorrect time", function () {
-				o(parseTimeString("12:3m")).equals(null)
-				o(parseTimeString("A:3")).equals(null)
-				o(parseTimeString("")).equals(null)
-				o(parseTimeString(":2")).equals(null)
-				o(parseTimeString("25:03")).equals(null)
-				o(parseTimeString("22:93")).equals(null)
-				o(parseTimeString("24")).equals(null)
-				o(parseTimeString("13pm")).equals(null)
-				o(parseTimeString("263PM")).equals(null)
-				o(parseTimeString("1403PM")).equals(null)
-				o(parseTimeString("14:03:33PM")).equals(null)
-				o(parseTimeString("9:37 acme")).equals(null)
-			})
-			o("parses AM/PM time", function () {
-				o(parseTimeString("7PM")).deepEquals({
-					hours: 19,
-					minutes: 0,
-				})
-				o(parseTimeString("11PM")).deepEquals({
-					hours: 23,
-					minutes: 0,
-				})
-				o(parseTimeString("12PM")).deepEquals({
-					hours: 12,
-					minutes: 0,
-				})
-				o(parseTimeString("11:30PM")).deepEquals({
-					hours: 23,
-					minutes: 30,
-				})
-				o(parseTimeString("12AM")).deepEquals({
-					hours: 0,
-					minutes: 0,
-				})
-				o(parseTimeString("12:30AM")).deepEquals({
-					hours: 0,
-					minutes: 30,
-				})
-				o(parseTimeString("3:30AM")).deepEquals({
-					hours: 3,
-					minutes: 30,
-				})
-				o(parseTimeString("3:30PM")).deepEquals({
-					hours: 15,
-					minutes: 30,
-				})
-				o(parseTimeString("9:37am")).deepEquals({
-					hours: 9,
-					minutes: 37,
-				})
-				o(parseTimeString("1:59pm")).deepEquals({
-					hours: 13,
-					minutes: 59,
-				})
-				o(parseTimeString("3:30 AM")).deepEquals({
-					hours: 3,
-					minutes: 30,
-				})
-				o(parseTimeString("3:30 PM")).deepEquals({
-					hours: 15,
-					minutes: 30,
-				})
-				o(parseTimeString("9:37 am")).deepEquals({
-					hours: 9,
-					minutes: 37,
-				})
-				o(parseTimeString("1:59 pm")).deepEquals({
-					hours: 13,
-					minutes: 59,
-				})
-				o(parseTimeString("9:37 a.m.")).deepEquals({
-					hours: 9,
-					minutes: 37,
-				})
-				o(parseTimeString("1:59 p.m.")).deepEquals({
-					hours: 13,
-					minutes: 59,
-				})
-				o(parseTimeString("1052 P.M.")).deepEquals({
-					hours: 22,
-					minutes: 52,
-				})
-				o(parseTimeString("1052 A.M.")).deepEquals({
-					hours: 10,
-					minutes: 52,
-				})
-				o(parseTimeString("948 P.M.")).deepEquals({
-					hours: 21,
-					minutes: 48,
-				})
-				o(parseTimeString("948 A.M.")).deepEquals({
-					hours: 9,
-					minutes: 48,
-				})
-			})
-		})
-		o.spec("timeDiff", function () {
-			o("A minor than B, with 15 min diff", function () {
-				const timeA = new Time(8, 35)
-				const timeB = new Time(8, 50)
-				o(timeA.diff(timeB)).equals(15)
-			})
-			o("A greater than B", function () {
-				const timeA = new Time(8, 50)
-				const timeB = new Time(8, 35)
-				o(timeA.diff(timeB)).equals(1425)
-			})
-			o("A minor than B, with one hour diff", function () {
-				const timeA = new Time(8, 0)
-				const timeB = new Time(9, 0)
-				o(timeA.diff(timeB)).equals(60)
-			})
-			o("diff with midnight", function () {
-				const timeA = new Time(23, 0)
-				const timeB = new Time(0, 0)
-				o(timeA.diff(timeB)).equals(60)
-			})
-			o("diff between two days - over midnight", function () {
-				const timeA = new Time(23, 0)
-				const timeB = new Time(1, 0)
-				o(timeA.diff(timeB)).equals(120)
-			})
-		})
-		o.spec("timeAdd", function () {
-			o("add 15 minutes", function () {
-				const timeA = new Time(8, 35)
-				const timeB = new Time(8, 50)
-				o(timeA.add({ minutes: 15 }).toObject()).deepEquals(timeB.toObject())
-			})
-			o("add 1 hours", function () {
-				const timeA = new Time(8, 35)
-				const timeB = new Time(9, 35)
-				o(timeA.add({ hours: 1 }).toObject()).deepEquals(timeB.toObject())
-			})
-			o("add 1 hour and 15 minutes", function () {
-				const timeA = new Time(8, 35)
-				const timeB = new Time(9, 50)
-				o(timeA.add({ hours: 1, minutes: 15 }).toObject()).deepEquals(timeB.toObject())
-			})
-			o("add 600 minutes overflowing to 'next day'", function () {
-				const timeA = new Time(14, 0)
-				const timeB = new Time(0, 0)
-				o(timeA.add({ minutes: 600 }).toObject()).deepEquals(timeB.toObject())
-			})
-			o("add 10 hours overflowing to 'next day'", function () {
-				const timeA = new Time(14, 0)
-				const timeB = new Time(0, 0)
-				o(timeA.add({ hours: 10 }).toObject()).deepEquals(timeB.toObject())
-			})
-			o("add 70 minutes and 11 hours overflowing to 'next day'", function () {
-				const timeA = new Time(14, 0)
-				const timeB = new Time(2, 10)
-				o(timeA.add({ hours: 11, minutes: 70 }).toObject()).deepEquals(timeB.toObject())
-			})
-		})
-		o.spec("timeSub", function () {
-			o("sub 15 minutes", function () {
-				const timeA = new Time(8, 35)
-				const timeB = new Time(8, 20)
-				o(timeA.sub({ minutes: 15 }).toObject()).deepEquals(timeB.toObject())
-			})
-			o("sub 30 minutes from minute 0", function () {
-				const timeA = new Time(8, 0)
-				const timeB = new Time(7, 30)
-				o(timeA.sub({ minutes: 30 }).toObject()).deepEquals(timeB.toObject())
-			})
-			o("sub 1 hours", function () {
-				const timeA = new Time(8, 35)
-				const timeB = new Time(7, 35)
-				o(timeA.sub({ hours: 1 }).toObject()).deepEquals(timeB.toObject())
-			})
-			o("sub 1 hour and 15 minutes", function () {
-				const timeA = new Time(8, 35)
-				const timeB = new Time(7, 20)
-				o(timeA.sub({ hours: 1, minutes: 15 }).toObject()).deepEquals(timeB.toObject())
-			})
-			o("sub 90 minutes", function () {
-				const timeA = new Time(8, 30)
-				const timeB = new Time(7, 0)
-				o(timeA.sub({ hours: 0, minutes: 90 }).toObject()).deepEquals(timeB.toObject())
-			})
-			o("sub 600 minutes overflowing to 'previous day'", function () {
-				const timeA = new Time(9, 0)
-				const timeB = new Time(23, 0)
-				o(timeA.sub({ minutes: 600 }).toObject()).deepEquals(timeB.toObject())
-			})
-			o("sub 10 hours overflowing to 'previous day'", function () {
-				const timeA = new Time(9, 0)
-				const timeB = new Time(23, 0)
-				o(timeA.sub({ hours: 10 }).toObject()).deepEquals(timeB.toObject())
-			})
-			o("sub 70 minutes and 11 hours overflowing to 'previous day'", function () {
-				const timeA = new Time(9, 0)
-				const timeB = new Time(20, 50)
-				o(timeA.sub({ hours: 11, minutes: 70 }).toObject()).deepEquals(timeB.toObject())
-			})
-		})
-		o.spec("compareTimes", function () {
-			o("A is before B (isBefore)", function () {
-				const timeA = new Time(9, 0)
-				const timeB = new Time(23, 0)
-				o(timeA.isBefore(timeB)).equals(true)
-			})
-			o("A is after B (isBefore)", function () {
-				const timeA = new Time(23, 0)
-				const timeB = new Time(9, 0)
-				o(timeA.isBefore(timeB)).equals(false)
-			})
-			o("A is before B (isAfter)", function () {
-				const timeA = new Time(9, 0)
-				const timeB = new Time(23, 0)
-				o(timeA.isAfter(timeB)).equals(false)
-			})
-			o("A is after B (isAfter)", function () {
-				const timeA = new Time(23, 0)
-				const timeB = new Time(9, 0)
-				o(timeA.isAfter(timeB)).equals(true)
-			})
-			o("A is equal B", function () {
-				const timeA = new Time(9, 0)
-				const timeB = new Time(9, 0)
-				o(timeA.isAfter(timeB)).equals(false)
-				o(timeA.isBefore(timeB)).equals(false)
-			})
-		})
-		o.spec("fromMinutes", function () {
-			o.test("negative minutes", function () {
-				const time = Time.fromMinutes(-65)
-
-				o.check(time.hour).equals(1)
-				o.check(time.minute).equals(5)
-			})
-
-			o.test("positive minutes", function () {
-				const time = Time.fromMinutes(65)
-
-				o.check(time.hour).equals(1)
-				o.check(time.minute).equals(5)
-			})
-
-			o.test("zero minutes", function () {
-				const time = Time.fromMinutes(0)
-
-				o.check(time.hour).equals(0)
-				o.check(time.minute).equals(0)
-			})
-		})
-
-		o.test("to24HourString", function () {
-			o(new Time(15, 25).to24HourString()).equals("15:25")
-			o(new Time(0, 0).to24HourString()).equals("00:00")
-			o(new Time(10, 55).to24HourString()).equals("10:55")
-			o(new Time(22, 55).to24HourString()).equals("22:55")
-		})
-
-		o.spec("to12HourString", function () {
-			o.test("with AM/PM suffix", function () {
-				o(new Time(15, 25).to12HourString(true)).equals("3:25 pm")
-				o(new Time(0, 0).to12HourString(true)).equals("12:00 am")
-				o(new Time(10, 55).to12HourString(true)).equals("10:55 am")
-				o(new Time(22, 55).to12HourString(true)).equals("10:55 pm")
-			})
-
-			o.test("without AM/PM suffix", function () {
-				o(new Time(15, 25).to12HourString(false)).equals("3:25")
-				o(new Time(0, 0).to12HourString(false)).equals("12:00")
-				o(new Time(10, 55).to12HourString(false)).equals("10:55")
-				o(new Time(22, 55).to12HourString(false)).equals("10:55")
-			})
 		})
 	})
 	o.spec("getStartOfWeek", function () {
